app/src/main/java/home/stanislavpoliakov/meet19_weather/presentation/presenter/BriefData.java
package home.stanislavpoliakov.meet19_weather.presentation.presenter;

public class BriefData {
    private double temperatureMin;
    private double temperatureMax;
    private long time;

    public double getTemperatureMin() {
        return temperatureMin;
    }

    public void setTemperatureMin(double temperatureMin) {
        this.temperatureMin = temperatureMin;
    }

    public double getTemperatureMax() {
        return temperatureMax;
    }

    public void setTemperatureMax(double temperatureMax) {
        this.temperatureMax = temperatureMax;
    }

    public long getTime() {
        return time;
    }

    public void setTime(long time) {
        this.time = time;
    }
}
